Refactor: DiagnosticEngine v2.0.0 — new checks, MIT license, CHANGELOG
- Replace Application static class with DiagnosticEngine instance class
- Split into partial classes: Cluster, Node, Storage, Qemu, Lxc, Common
- Add new checks: ThinOvercommit, DiskCache, SnapshotCount, SnapshotAge,
  HealthScore, LXC Nesting/Keyctl, CPU x86-64 compatibility, VM locked,
  Unused disks for LXC, CPU hotplug Windows, Balloon, SecureBoot Win11
- Move IsLocked/OnBoot/Protection checks to CheckCommonVmAsync (shared)
- Add SettingsSnapshot and SettingsHealthScore settings sections
- Change license from GPL-3.0 to MIT
- Add CHANGELOG.md with use-changelog enabled in publish workflow
- Update README with full check reference and settings documentation
