+++ Directory.Packages.props
@@ -4,7 +4,7 @@
     <PackageOutputPath>..\nupkgs\</PackageOutputPath>
   </PropertyGroup>
   <ItemGroup>
-    <PackageVersion Include="Corsinvest.ProxmoxVE.Api.Extension" Version="9.1.5" />
-    <PackageVersion Include="Corsinvest.ProxmoxVE.Api.Console" Version="9.1.5" />
+    <PackageVersion Include="Corsinvest.ProxmoxVE.Api.Extension" Version="9.1.8" />
+    <PackageVersion Include="Corsinvest.ProxmoxVE.Api.Console" Version="9.1.8" />
   </ItemGroup>
 </Project>
