+++ src/Corsinvest.ProxmoxVE.Diagnostic.Api/DiagnosticResult.cs
@@ -1,6 +1,6 @@
 /*
  * SPDX-FileCopyrightText: Copyright Corsinvest Srl
- * SPDX-License-Identifier: GPL-3.0-only
+ * SPDX-License-Identifier: MIT
  */
 
 using System.Text.RegularExpressions;
