+++ src/Corsinvest.ProxmoxVE.Diagnostic.Api/DiagnosticResultContext.cs
@@ -1,6 +1,6 @@
 /*
  * SPDX-FileCopyrightText: Copyright Corsinvest Srl
- * SPDX-License-Identifier: GPL-3.0-only
+ * SPDX-License-Identifier: MIT
  */
 
 namespace Corsinvest.ProxmoxVE.Diagnostic.Api;
