+++ src/Corsinvest.ProxmoxVE.Diagnostic/Corsinvest.ProxmoxVE.Diagnostic.csproj
@@ -12,9 +12,9 @@
     <RepositoryType>git</RepositoryType>
   </PropertyGroup>
 
-  <ItemGroup>
+  <ItemGroup >
     <PackageReference Include="Corsinvest.ProxmoxVE.Api.Console" />
-
+    <!-- <ProjectReference Include="..\..\..\cv4pve-api-dotnet\src\Corsinvest.ProxmoxVE.Api.Console\Corsinvest.ProxmoxVE.Api.Console.csproj" /> -->
     <ProjectReference Include="..\Corsinvest.ProxmoxVE.Diagnostic.Api\Corsinvest.ProxmoxVE.Diagnostic.Api.csproj" />
   </ItemGroup>
 </Project>
